requirements.txt
streamlit
plotly
pandas
numpy
websockets
xgboost
python-telegram-bot
requests